+++ uniffi_bindgen/src/bindings/swift/templates/wrapper.swift
@@ -18,6 +18,7 @@ import {{ config.ffi_module_name() }}
 
 {% include "RustBufferTemplate.swift" %}
 {% include "Helpers.swift" %}
+{% include "HandleMap.swift" %}
 
 // Public interface members begin here.
 {{ type_helper_code }}